create main, top-level tests
