+++ tests/basic--impl_Debug_for_i32.rs
