+++ tests/basic--supertraits-hr.rs
