+++ tests/coinductive.rs
