+++ tests/impl-vs-trait-fn.rs
