+++ tests/wf-impl--supertrait-required.rs
